Git - test minimum version
